adding dashboard favorites UI
Former-commit-id: bc6927a61bd2efca616d454aeb079fb708c80cc3
Former-commit-id: 6e063492dc789bbfc5cd5d935d93b2bae3f91c78

+++ web/concrete/src/Package/PackageService.php
@@ -183,11 +183,19 @@ class PackageService
         }
         try {
 
-            $config = $this->entityManager->getConfiguration();
-            $driver = $config->getMetadataDriverImpl();
-            $driver->addPaths($p->getPackageEntityPaths());
-            $cache = $config->getMetadataCacheImpl();
-            $cache->flushAll();
+            $paths = array();
+            foreach($p->getPackageEntityPaths() as $path) {
+                if (is_dir($path)) {
+                    $paths[] = $path;
+                }
+            }
+            if (count($paths)) {
+                $config = $this->entityManager->getConfiguration();
+                $driver = $config->getMetadataDriverImpl();
+                $driver->addPaths($paths);
+                $cache = $config->getMetadataCacheImpl();
+                $cache->flushAll();
+            }
 
             $u = new \User();
             $swapper = new ContentSwapper();
